+++ src/main/java/webvet/v1/domain/aggregates/model/Cliente.java
@@ -3,7 +3,7 @@ package webvet.v1.domain.aggregates.model;
 public class Cliente {
 
 
-    private Integer clienteId;
+    private Long clienteId;
 
     private String nombre;
 
@@ -17,11 +17,11 @@ public class Cliente {
 
     private String ciudad;
 
-    public Integer getClienteId() {
+    public Long getClienteId() {
         return clienteId;
     }
 
-    public void setClienteId(Integer clienteId) {
+    public void setClienteId(Long clienteId) {
         this.clienteId = clienteId;
     }
